Promijenjen int u long kod broja lajkova kod objave

+++ app/src/main/java/hr/example/treeapp/Post.java
@@ -8,9 +8,10 @@ public class Post {
     double longitute;
     String Opis;
     String URL_slike;
-    int Broj_lajkova;
+    long Broj_lajkova;
 
-    public Post(String ID_objava, String korisnik_ID, String datum_objave, double latitude, double longitute, String opis, String URL_slike, int broj_lajkova) {
+
+    public Post(String ID_objava, String korisnik_ID, String datum_objave, double latitude, double longitute, String opis, String URL_slike, long broj_lajkova) {
         this.ID_objava = ID_objava;
         Korisnik_ID = korisnik_ID;
         Datum_objave = datum_objave;
@@ -21,11 +22,12 @@ public class Post {
         Broj_lajkova = broj_lajkova;
     }
 
-    public int getBroj_lajkova() {
+
+    public long getBroj_lajkova() {
         return Broj_lajkova;
     }
 
-    public void setBroj_lajkova(int broj_lajkova) {
+    public void setBroj_lajkova(long broj_lajkova) {
         Broj_lajkova = broj_lajkova;
     }
